Create New 'Employee' App. goal: Decouple from 'Hardware' App
The initial phase of the project had Employee & Hardware (Laptop) (and all
their related models) in one app: 'Hardware' but as we move further in
development we will add more 'Employee' specific functionalities & features.

Keeping 'Employee' coupled in the 'Hardware' app will make it complicated to
manage.

Given the current number of employee views, forms etc, and the future
plans (GSuite User Generation, Detail Employee Self-service/Profile etc)
It is best to move the model to it's own app.