Separate employee related functionalities into it's own app (#3)
* Update Readme with more information

* Update Tests/CI Badge link (Remove actions from .svg file)

* Fix Build Badge Link

* Create New 'Employee' App. goal: Decouple from 'Hardware' App

The initial phase of the project had Employee & Hardware (Laptop) (and all
their related models) in one app: 'Hardware' but as we move further in
development we will add more 'Employee' specific functionalities & features.

Keeping 'Employee' coupled in the 'Hardware' app will make it complicated to
manage.

Given the current number of employee views, forms etc, and the future
plans (GSuite User Generation, Detail Employee Self-service/Profile etc)
It is best to move the model to it's own app.

* Add 'Employee' app to settings

* Re-link all Employee (and related) model references to new app

With the 'Employee' (and related) models decoupled from the hardware
app into it's own app, all existing referneces to the models have to be
edited.

* Move 'Employee' (and related) Models to 'employee' app

Move the Model classes to the employee app and prepare for new migrations

* Add Migrations

* Decouple/Move URL & Views code to respective apps

With us now having separate apps for Employee related functions & Laptop
related Functions, the URLs & Views should be moved to their respective
apps (currently all urls & views in Hardware app) to keep code logically
placed in project.

Common urls & views (login, dashboard etc) moved to project directory for
the above reason. These urls & views don't belong to any one specific app
and thus the project dir would be the best place for the same.

The URLs have been updated for this change, all 'Employee' urls are prefixed
with 'employee/' and hardware urls with 'hardware'. This keeps the URls
clean and realated to the respective app views.

* Move User Creation signal code to employee app

* Move serializer from hardware app to employee app

* Move employee related forms to employee app from hardware app

* Move employee filters from employee app to hardware app

* Move view decorators to project dir (minierp)

As all current decorators are common for both employee and hardware app,
we should move them to the project dir. This will keep better code placement
logic.

The views have to be updated to read from the new decorators location.

* Move employee related admin.py code to employee app

* Move tests directory to employee app

* Fix Bugs from Test Result

Bugs:
- Wrong source of decorators in minierp/views (was referencing hardware)
- load_building in employee views instead of hardware views
- Order of url patterns checking for employee detail page pattern when
running test for employee add (first checks 'employee/x' pattern > found
x should be an pk, but request was x to be add/, changed the order to
fix this)
- employee/test_views EmployeeDashboard test checking base url ('employee')
instead of dashboard ('/')

* Moved templates folder to root directory & re-order url patterns

- Move the templates folder from hardware app dir to main project dir. This
keeps things tidy as similar to how we have a single project static folder
(with app sub folders) we will also now have a single project templates
folder with app based templates

The templates location should also be updated on all child templates of
main.html to maintain inheritance.

= URL patterns have to be updated to fix conflict/mis-match: As url patterns
are checked sequentially, the first pattern is taken as the request url
pattern. This means employee/x will taken the employee/<str:pk> pattern
instead of employee/add unless re-ordered.

[ ] A better solution needs to be worked on for the above (Point 3)

* Update views & templates for the new templates directory change

* Fix Bugs from Test Results