Add & Configure new Finance app
Departments in every organization deals with finance be it from paying
monthly bills, coordinating with the Finance department for fund allocation
or forecasting future payments. Finance is a core area of any department.

While payments (making payments) cannot be automate (atleast not in the
current scope), administration/management of payments can be automated
to an extend.

This is what this Finance app aims to do. Automating Reminders, forecasting
future payments, Sending mails to finance for fund allocation, managing
invoices are some of the items in scope.