Create Custom Template Tag for Currency Representation
All cost realated fields are IntergerFields. As such, them rendering
on the template does not provide the most accurate representation.

Using custom template tags that make use of the 'locale' python module,
these numbers are taken and converted into the local currency with decimal,
commas & symbol and then rendered on the template.

1000 > ₹ 1,000.00