Update DRF Searilzers to use PrimaryKeyRelatedID for allowing POST
This change will allow DRF Endpoints to allow workable POST requests.
With StringRelatedID (SRI), the payload was very human readable but did not
allow POSTing data as SRIs are read_only + there's no ID making it hard
to work with.

With PrimaryKeyRealatedID, this problem gets solved. We have the id to
query & key plus it's value to render.