Handle OperationalError Raised When Running CI/CD Actions
- The OperationalError is raised due to referencing the new EmployeeAppSetting
& HardwareAppSetting models inside views, forms & signals. These references
must exist/be added only after full migration. Any operation being run
(migrations, tests, runserver) will check for these references, if not found
(which it won't as we have not migrated (correction: we cannot)), an
OperationalError is raised.
- In development/production, this was solved by 1. removing all the
reference to the unmigrated table(s) 2. migrating & 3. Adding in back
the references.
- For Actions, this cannot be done as every run is afresh. We remove references,
it works. We add it back in, OperationalError is raised as the tables do not
exist.
- To solve it ONLY for Actions, we handle the OperationalError raised by
catching it and 'pass'ing it.

references: https://stackoverflow.com/a/36503812/17408451