Separate HardwareAppSetting Model Migration into it's own migration file
- HardwareAppSetting is a model that has not yet been migrated into the DB.
As expected when migrations are generated, it has the CreateModel
method for HardwareAppSetting but the challenge is when reset the migrations.

Resetting the migrations invole 'faking' the migrations (because the
tables already exist (we presume) and don't have to actually be migrated)

With HardwareAppSetting (unmigrated) added along the others in the migration
file, the model is never migrated due to faked migration.

To solve this, the HardwareAppSetting migration has been moved to 2nd
empty migration and will be run as an actual migration after the faked
migration of 0001